feat: Implemented my own game images, style and fonts
It was implemented my own set of colors and fonts to the game. It was
also implemented rock style and images and the shuffle functionality.

As now, games has an array of 16 card elements with rock images. But
no click action is added to it, meaning that nothing can be done with
the game yet.